moved GameOver to components folder
