+++ src/components/index.ts
@@ -1,5 +1,6 @@
-import Loading from './Loading';
+import GameOver from './GameOver';
 import Header from './Header';
+import Loading from './Loading';
 import MovieCard from './MovieCard';
 
-export { Loading, Header, MovieCard };
+export { GameOver, Header, Loading, MovieCard };
