+++ src/pages/index.tsx
@@ -1,6 +1,5 @@
 import Home from './Home';
 import Game from './Game';
-import GameOver from './GameOver';
 import ScoreBoard from './ScoreBoard';
 
-export { Home, Game, GameOver, ScoreBoard };
+export { Home, Game, ScoreBoard };
